Acrescentado o conftest.py
close #16
